+++ lib/ankos/doubles_anko.dart
@@ -1,15 +1,28 @@
+import 'dart:async';
+
 import 'package:logging/logging.dart';
 
 extension DoubleAnko on double {
-  void log({String? tag}) {
-    Logger(tag ?? 'TAG').log(Level.INFO, this);
+  void log({String? tag, Object? error, StackTrace? stackTrace, Zone? zone}) {
+    Logger(tag ?? 'TAG').log(Level.INFO, this, error, stackTrace, zone);
   }
 
-  void logError({String? tag}) {
-    Logger(tag ?? 'TAG').log(Level.SEVERE, this);
+  void logError(
+      {String? tag, Object? error, StackTrace? stackTrace, Zone? zone}) {
+    Logger(tag ?? 'TAG').log(Level.SEVERE, this, error, stackTrace, zone);
   }
 
-  void logWarning({String? tag}) {
-    Logger(tag ?? 'TAG').log(Level.WARNING, this);
+  void logWarning(
+      {String? tag, Object? error, StackTrace? stackTrace, Zone? zone}) {
+    Logger(tag ?? 'TAG').log(Level.WARNING, this, error, stackTrace, zone);
+  }
+
+  void logOther(
+      {String? tag,
+      required Level level,
+      Object? error,
+      StackTrace? stackTrace,
+      Zone? zone}) {
+    Logger(tag ?? 'TAG').log(level, this, error, stackTrace, zone);
   }
 }